@@ -9,7 +9,7 @@ describe('constructor', () => {
     });
 
     it('should throw if bad buffer passed', () => {
-        assert.throws(() => new PngImg(new Buffer({})));
+        assert.throws(() => new PngImg(Buffer.alloc(0)));
     });
 
     it('should not throw with valid image passed', () => {
